requirements_coinenv.txt
# conda
python-dotenv
openai
PyYAML

# pip
pyupbit
feedparser # for rss
bs4 # for rss
ta # Widely used library for technical indicators: https://github.com/bukosabino/ta?tab=readme-ov-file
openpyxl
youtube-transcript-api
--upgrade ta # library for calculating techinacal index automatically
selenium
webdriver-manager
Pillow
langchain_openai
-U vllm
gguf>=0.13.0,<0.18
streamlit
plotly
httpx
bitsandbytes>=0.46.1